requirements.txt
asgiref==3.7.2
Django==5.0
django-tinymce==3.7.1
sqlparse==0.4.4
typing_extensions==4.8.0
